实现部分自定义模块功能

+++ service/service-customize/src/main/java/org/sunt/customize/meta/DataType.kt
@@ -2,8 +2,8 @@ package org.sunt.customize.meta
 
 enum class DataType(val label: String) {
 
-    INTEGER("Int"),
-    DOUBLE("Double"),
+    INTEGER("Integer"),
+    DECIMAL("Decimal"),
     STRING("String"),
     BOOLEAN("Boolean"),
     DATE("Date"),
